rewrite syncstring_schema part of db-schema in typescript
- this involved a lot of little tweaks due to stuff discovered by
typescript.
- probably no real changes in the code though
