+++ src/packages/util/schema.ts
@@ -12,9 +12,6 @@ export { RECENT_TIMES, RECENT_TIMES_KEY } from "./schema-static";
 
 export { SCHEMA, client_db, site_settings_conf } from "./db-schema";
 
-// Load the syncstring extensions to the schema
-import "./syncstring_schema";
-
 // Will import some other modules and make them available here, since the code
 // used to be in this file, and this is assumed in code elsewhere.  Will change later.
 export { COMPUTE_STATES } from "./compute-states";
